Use pydantic to validate messages
