+++ tasks/loader/callback.py
@@ -1,14 +1,17 @@
-import json
+import logging
 import traceback
 from typing import Awaitable, Callable, List
 
 from opentelemetry import trace
 from opentelemetry.trace import SpanKind
 from opentelemetry.trace.propagation.tracecontext import TraceContextTextMapPropagator
+from pydantic import ValidationError
 
 from common.nats import Msg
 
 from .types import Event, Handler
+
+logger = logging.getLogger(__name__)
 
 tracer = trace.get_tracer(__name__)
 propagator = TraceContextTextMapPropagator()
@@ -39,14 +42,28 @@ def generate(
                 "task.handlers.total": len(handlers),
             },
         ) as span:
-            with tracer.start_as_current_span("parse"):
-                kwargs = json.loads(message.data)
+
+            # Parse and validate the input
+            try:
+                with tracer.start_as_current_span("parse"):
+                    kwargs = event.input_validator.parse_raw(
+                        message.data.decode("utf-8")
+                    )
+            except ValidationError:
+                ctx = span.get_span_context()
+                logger.error(
+                    f"invalid input for {event.name} ({event.KIND}) - trace id: {ctx.trace_id}"
+                )
+
+                # Acknowledge the message on failure since retrying it won't yield any benefits
+                await message.ack()
+                return
 
             failed = 0
             for handler in handlers:
                 try:
                     with tracer.start_as_current_span(handler.name):
-                        await handler.callback(**kwargs)
+                        await handler.callback(**kwargs.dict())
                 except Exception as e:
                     failed += 1
                     traceback.print_exception(e)  # type: ignore
