+++ tasks/loader/types/automated/registration.py
@@ -1,0 +1,9 @@
+from enum import auto
+
+from .base_action import BaseAction
+
+
+class Registration(BaseAction):
+    NewApplication = auto()
+    Accepted = auto()
+    Rejected = auto()
