+++ tasks/loader/types/automated/sync.py
@@ -1,0 +1,7 @@
+from enum import auto
+
+from .base_action import BaseAction
+
+
+class Sync(BaseAction):
+    SignUp = auto()
