+++ tasks/loader/types/automated/unused_actions.py
@@ -1,0 +1,13 @@
+from .base_action import BaseAction
+
+
+class Communication(BaseAction):
+    pass
+
+
+class Integrations(BaseAction):
+    pass
+
+
+class Workshops(BaseAction):
+    pass
