+++ tasks/loader/types/errors.py
@@ -2,30 +2,6 @@ class LoaderException(Exception):
     pass
 
 
-class InvalidEvent(LoaderException):
-    """
-    An error occurred while parsing the event
-    """
-
-    pass
-
-
-class UnknownService(InvalidEvent):
-    """
-    The specified service does not exist
-    """
-
-    pass
-
-
-class UnknownAction(InvalidEvent):
-    """
-    The specified action does not exist on the service
-    """
-
-    pass
-
-
 class InvalidCallback(LoaderException):
     """
     An error occurred while validating the callback
